api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openpyxl==3.1.2
celery==5.3.6
redis==5.0.1
pytest==8.0.0
black==24.1.1
isort==5.13.2
python-dotenv==1.0.1
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic[email]==2.6.1
alembic==1.13.1
boto3==1.34.29
resume-parser==0.8.4
skillner==1.0.3
openai==1.10.0
supabase==2.3.4
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 